unity builds, remove nlohmann

+++ src/hacks/Global/DiscordRPC.cpp
@@ -296,7 +296,7 @@ namespace eclipse::hacks::Global {
 
         void update() override {
             if (!config::get<"global.discordrpc", bool>(false)) return;
-            auto interval = config::get<"global.discordrpc.interval", float>(200.0f);
+            auto interval = config::get<"global.discordrpc.interval", double>(200.0f);
             auto now = std::chrono::steady_clock::now();
             auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(now - s_lastDiscordUpdate).count();
             if (static_cast<float>(elapsed) >= interval) {
