+++ src/hacks/Global/Speedhack.cpp
@@ -22,7 +22,7 @@ namespace eclipse::hacks::Global {
 
         [[nodiscard]] bool isCheating() const override {
             auto toggle = config::get<"global.speedhack.toggle", bool>();
-            auto speed = config::get<"global.speedhack", float>(1.f);
+            auto speed = config::get<"global.speedhack", double>(1.f);
             return toggle && speed != 1.f;
         }
 
@@ -36,7 +36,7 @@ namespace eclipse::hacks::Global {
         ADD_HOOKS_DELEGATE("global.speedhack.toggle")
 
         void update(float dt) override {
-            auto speed = config::get<float>("global.speedhack", 1.f);
+            auto speed = config::get<double>("global.speedhack", 1.f);
 
             if (speed <= 0)
                 speed = 1.f;
